break TransactionSummary and ShippingAndPaymentDetails apart
