+++ src/Apps/Order/Routes/Offer/index.tsx
@@ -1,8 +1,9 @@
 import { Button, Flex, Message, Sans, Spacer } from "@artsy/palette"
 import { Offer_order } from "__generated__/Offer_order.graphql"
 import { OfferMutation } from "__generated__/OfferMutation.graphql"
+import { ArtworkSummaryItemFragmentContainer as ArtworkSummaryItem } from "Apps/Order/Components/ArtworkSummaryItem"
 import { Helper } from "Apps/Order/Components/Helper"
-import { TransactionSummaryFragmentContainer as TransactionSummary } from "Apps/Order/Components/TransactionSummary"
+import { TransactionDetailsSummaryItemFragmentContainer as TransactionDetailsSummaryItem } from "Apps/Order/Components/TransactionDetailsSummaryItem"
 import { TwoColumnLayout } from "Apps/Order/Components/TwoColumnLayout"
 import { ContextConsumer, Mediator } from "Artsy/SystemContext"
 import { Input } from "Components/Input"
@@ -188,18 +189,21 @@ export class OfferRoute extends Component<OfferProps, OfferState> {
             }
             Sidebar={
               <Flex flexDirection="column">
-                <TransactionSummary
-                  order={order}
-                  mb={[2, 3]}
-                  offerOverride={
-                    this.state.offerValue &&
-                    this.state.offerValue.toLocaleString("en-US", {
-                      style: "currency",
-                      currency: "USD",
-                      minimumFractionDigits: 2,
-                    })
-                  }
-                />
+                <Flex flexDirection="column">
+                  <ArtworkSummaryItem order={order} />
+                  <TransactionDetailsSummaryItem
+                    order={order}
+                    offerOverride={
+                      this.state.offerValue &&
+                      this.state.offerValue.toLocaleString("en-US", {
+                        style: "currency",
+                        currency: "USD",
+                        minimumFractionDigits: 2,
+                      })
+                    }
+                  />
+                </Flex>
+                <Spacer mb={[2, 3]} />
                 <Helper artworkId={artwork.id} />
                 <Media at="xs">
                   <>
@@ -257,7 +261,8 @@ export const OfferFragmentContainer = createFragmentContainer(
           }
         }
       }
-      ...TransactionSummary_order
+      ...ArtworkSummaryItem_order
+      ...TransactionDetailsSummaryItem_order
     }
   `
 )
