+++ src/Apps/__tests__/Fixtures/Order.ts
@@ -56,7 +56,7 @@ export const UntouchedBuyOrder = {
                 url:
                   "https://d7hftxdivxxvm.cloudfront.net?resize_to=fit&width=185&height=184&quality=80&src=https%3A%2F%2Fd32dm0rphc51dk.cloudfront.net%2FtOfWds4sIX_9WpRf3RqaQQ%2Flarge.jpg",
               },
-              resized_transactionSummary: {
+              resized_ArtworkSummaryItem: {
                 url:
                   "https://d7hftxdivxxvm.cloudfront.net?resize_to=fit&width=185&height=184&quality=80&src=https%3A%2F%2Fd32dm0rphc51dk.cloudfront.net%2FtOfWds4sIX_9WpRf3RqaQQ%2Flarge.jpg",
               },
