+++ src/Utils/ExtractProps.ts
@@ -1,0 +1,3 @@
+export type ExtractProps<
+  T extends React.ComponentType<any>
+> = T extends React.ComponentType<infer Props> ? Props : never
